bot-jhon/requirements.txt
discord.py>=2.3.2
yt-dlp>=2024.0.0
PyNaCl>=1.5.0
python-dotenv>=1.0.0
requests>=2.31.0
qrcode[pil]>=7.4.2
asyncio
